requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
openai==0.28.1
Werkzeug==2.3.7
gunicorn==21.2.0
psycopg2-binary==2.9.7
speechrecognition==3.10.0
pydub==0.25.1
googletrans==4.0.0
requests==2.31.0
celery==5.3.4
redis==5.0.1
